+++ src/components/ConsortiaControlledVocabulary/modals/index.js
@@ -1,3 +1,4 @@
+export { ConfirmCreateEntryModal } from './ConfirmCreateEntryModal';
 export { ConfirmDeleteEntryModal } from './ConfirmDeleteEntryModal';
 export { ConfirmShareEntryModal } from './ConfirmShareEntryModal';
 export { ItemInUseModal } from './ItemInUseModal';