+++ src/routes/ConsortiumManager/settings/users/general/Departments/validate.js
@@ -1,32 +1,14 @@
 import { FormattedMessage } from 'react-intl';
 
-import {
-  validateRequired,
-  validateUniqueness,
-} from '../../../../../../components/ConsortiaControlledVocabulary/validators';
+import { validateRequired } from '../../../../../../components/ConsortiaControlledVocabulary/validators';
 
-export const validate = (item, index, items, initialValues) => {
-  const validateFieldUniqueness = (field, message) => validateUniqueness({
-    index,
-    item,
-    items,
-    field,
-    message,
-    initialValues,
+export const validate = (item) => {
+  const codeError = validateRequired({
+    value: item.code,
+    message: <FormattedMessage id="ui-users.settings.departments.code.required" />,
   });
 
-  const nameError = validateFieldUniqueness('name', <FormattedMessage id="ui-users.settings.departments.name.error" />);
-
-  const codeError = (
-    validateRequired({
-      value: item.code,
-      message: <FormattedMessage id="ui-users.settings.departments.code.required" />,
-    })
-    || validateFieldUniqueness('code', <FormattedMessage id="ui-users.settings.departments.code.error" />)
-  );
-
   return {
-    name: nameError,
     code: codeError,
   };
 };